support marking the original data as pruned
during pruning, original data is re-saved to (<version>, 0)
and the original data is deleted. this allows the data to still be
referenced by ReferenceNodes saved at future versions, but without
affecting the earliest version recognized by the tree.

this --prune-node flag does this with the root node, resulting in a tree
state that has the correct data referenced at the desired version, but
no recognized data stored at the original version.

this is useful for fixing data that has saved the root node to version 1.