chore: Add hardhat e2e test suite for precompiles
This adds a tests/e2e-evm hardhat based e2e test suite for testing EVM
and EVM<>Cosmos workflows.  This will be used to drive and test the
implementation of precompiles with a focus on e2e workflows.

The hardhat and kvtool networks are configured to run in CI to ensure
continuous support for both networks is not broken.  The hardhat network
is used to ensure compatibility between Kava and other EVM's, as well as
offer better support for unit testing supporting and example contracts.

The kvtool network runs a Kava network for testing the EVM
implementation on Kava and testing stateful precompiles.