chore: Update npm test commands
This adds a test-hardhat and test-kvtool command that selects the proper
network and runs typechecks.

In addition, npm test is updated to run test-hardhat, then test-kvtool.
The hardhat tests are run first to fail fast, and the test-kvtool tests
are run after as tests that pass hardhat should pass on kvtool.
However, kvtool will target additional tests that will not be run on the
hardhat network.