chore: Finish e2e-evm continuous integration
This refactors the e2e-evm tests to run in a single job with the hardhat
network running first to fail fast and trigger any type errors.  In
addition, contracts used to test against kvtool may be unit testing only
in the hardhat network, etc.

In addition, a curl check is added to ensure the EVM JSON-RPC is serving
requests.  This endpoint can take ~30 seconds to come online after the
kvtool server starts.  Curl is configured to check every 5 seconds and
retry 12 times over 60 seconds if the first request is not successful.

Linting and formatting checks are also added via eslint,
typescript-eslint with advanced type check rules, and prettier.

Solhint is installed, but errors without any files so is not configured
to run in CI yet.