feat: Extend noop precompile testing addresses
We extend the total registered precompile addresses to 7 with six of
those enabled for testing against the e2e-evm test suite.

This precompile addresses will form the basis for ensuring we implement
abi basic calls correctly and there are no regressions as more complex
abi behavior is added and tested.