test: Add contract call, gas, and balance tests
This adds two contract that call the Noop set of contracts.  The first
is the low level caller with uses opcodes directoy and supports Call,
DelegateCall, and StaticCall. Callcode to be implemented later.

In addition, a high level caller that calls the functions directly from
the interface, which includes a extcodesize check when compiled to
ensure that path is tested.

In addition, gas exhaustion checks are added to ensure expected reverts
still refund gas (ErrExecutionReverted), and balance checks before and
after transfers are added to ensure balance changes happen and revert as
expected.