test: Add E2E tests for a disabled precompile
Adds a mock with the same revert reason for all methods in the noop
interface plus receieve and fallback.  When called this mock and the
precompile will both revert and provide a reason.

External call, message call, staticall, and delegate call are tested to
revert when disabled.  Callcode to be added later.

Even though delegatecall may be disabled in release, we still want to
reserve this for individual stateful precompiles and allow the disabled
check to take precedence.