refactor: Dynamic generation basic ABI test cases
Changes from including a field in each test case from conditionally
running the case, to building the cases dynamically. This allows for
logical grouping of test cases and organization with logic instead of
using comments. Slightly less explicit for each test case, but with the
grouping of test cases, it reduces the mental overhead of figuring out
when each test case is run.