fix: Prevent assertion skip on 0 expected value
Expecting 0 value would make the conditional fail with a falsey value.
Adding additional else case to ensure the return value is undefined as
an extra check in case the assertions are not being run.

This correctly validates msg.value does indeed have the value passed to
callcode(), just not the same one as msg.value of the parent caller.