refactor: Test msg sender, value for all msg calls
This restructures the tests to confirm the expected msg.value and
msg.sender with various different message call types, e.g. call,
delegatecall, callcode, staticcall.

The storage tests are also split up so that they are independent set of
tests to keep different types of tests separate and reduce clutter with
the slight increase in redundancy in the test logic.