fix: Increase ABI basic test contractCallerGas
With the previous changes of functionCallCode calling the
functionCallCodeWithValue cause an increase in gas usage since calling
an function marked as external consume more gas in addition to the
overhead of calling anothher function.

This function was added to allow more control over the value the caller
contract sends with the message call to the implementation contract.