test: Expand ctx/storage tests for precompiles
This implements shared ABI interfaces for both the mock contracts and
later on, precompiles, that are used for the tests.

The main goal is to test precompiles handle different message call types
correctly, msg.sender, msg.value, storage location.

This expands the current set of tests that simply define the behavior of
call message types to a set of tests that can also enforce the same
behavior on a set of precompiles.