fix!: bump ethermint to fix EIP712 message signing
Previous version of ethermint used a value for the `verifyingContract`
in the domain separator that was not compliant with the spec. This field
and `salt`, which was unused, have been removed.

This is a state breaking change because it changes which EIP712 txs are
considered valid, and therefore, which txs are allowed to be included in
a block.