refactor(tests/e2e-evm): Restructure ABI tests for extensibility & reduce duplication  (#2031)
* refactor: Initial restructure of ABI basic function tests

* refactor: Restructure of ABI fallback tests

* fix: Use separate shouldRun fn signatures for fallback tests

Fallback tests are not run on each function but rather only for the specific fallback.
This means whether or not to run the test is determined by only the receive and fallback
functions, without any other ABI function.

* fix: Resolve use of type assertions, enforcing types

Previously used type assertions to bypass certain TypeScript issues with
test cases, along with using any & unsafe assignments. This resolves the
types to be properly valid and enforced to prevent any potential errors.

* refactor: Rename unused params in test cases

* refactor: Dynamic generation basic ABI test cases

Changes from including a field in each test case from conditionally
running the case, to building the cases dynamically. This allows for
logical grouping of test cases and organization with logic instead of
using comments. Slightly less explicit for each test case, but with the
grouping of test cases, it reduces the mental overhead of figuring out
when each test case is run.

* refactor: Resolve ABI disabled types & errors

Resolves use type casting and unsafe access, validation of revert errors
for matches and types

* refactor: Only test receive,  fallback funcs when available

Previously runs all the time, which is currently okay with the current
single testing contract that includes both functions. This conditionally
adds these test cases if the respective functions exist so we can test
additional contract behavior that may not have these functions and may
produce a different error.

* lint: Ignore solhint issues for mock contracts

Most of these issues are intentional and are okay to ignore. This also
sets the solhint ignoreConstructors option to true for the
func-visibility rule, as we are using solidity >=0.7.0

* refactor: Remove unused empty ContractTestCase generic brackets

* chore: Disable chai assertion error truncation

Long errors are truncated and difficult to determine the issue otherwise